fix balancer fetcher failing when file does not exists

+++ Agave/BalancerLiquidityFetcher.js
@@ -17,8 +17,11 @@ const gasPrice = BigNumber.from(0);
  * @param {string} rpcUrl 
  */
 async function FetchBalancerLiquidity(aave, rpcUrl) {
-    console.log('Deleting old balancer_volume_for_slippage.json file');
-    fs.rmSync('balancer_volume_for_slippage.json');
+    if(fs.existsSync('balancer_volume_for_slippage.json')) {
+        console.log('Deleting old balancer_volume_for_slippage.json file');
+        fs.rmSync('balancer_volume_for_slippage.json');
+    }
+
     const balancer = new BalancerSDK({
         network: 100, // gnosis
         rpcUrl: rpcUrl,
